refactor: remove backend calls in the booking component
